+++ pkg/toolset/p2p_identity_gen.go
@@ -6,28 +6,29 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/libp2p/go-libp2p-core/crypto"
 	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/mr-tron/base58"
 
-	"github.com/iotaledger/hive.go/configuration"
-
 	p2pCore "github.com/gohornet/hornet/core/p2p"
 	"github.com/gohornet/hornet/pkg/p2p"
+	"github.com/iotaledger/hive.go/configuration"
 )
 
-func extractP2PIdentity(nodeConfig *configuration.Configuration, args []string) error {
+func generateP2PIdentity(nodeConfig *configuration.Configuration, args []string) error {
+
 	printUsage := func() {
 		println("Usage:")
-		println(fmt.Sprintf("	%s [P2P_DATABASE_PATH]", ToolP2PExtractIdentity))
+		println(fmt.Sprintf("	%s [P2P_DATABASE_PATH]", ToolP2PIdentityGen))
 		println()
 		println("	[P2P_DATABASE_PATH] - the path to the p2p database folder (optional)")
 		println()
-		println(fmt.Sprintf("example: %s %s", ToolP2PExtractIdentity, "p2pstore"))
+		println(fmt.Sprintf("example: %s %s", ToolP2PIdentityGen, "p2pstore"))
 	}
 
 	if len(args) > 1 {
 		printUsage()
-		return fmt.Errorf("too many arguments for '%s'", ToolP2PExtractIdentity)
+		return fmt.Errorf("too many arguments for '%s'", ToolP2PIdentityGen)
 	}
 
 	p2pDatabasePath := nodeConfig.String(p2pCore.CfgP2PDatabasePath)
@@ -36,37 +37,47 @@ func extractP2PIdentity(nodeConfig *configuration.Configuration, args []string) 
 	}
 	privKeyFilePath := filepath.Join(p2pDatabasePath, p2p.PrivKeyFileName)
 
+	if err := os.MkdirAll(p2pDatabasePath, 0700); err != nil {
+		return fmt.Errorf("could not create peer store database dir '%s': %w", p2pDatabasePath, err)
+	}
+
 	_, err := os.Stat(privKeyFilePath)
 	switch {
-	case os.IsNotExist(err):
-		// private key does not exist
-		return fmt.Errorf("private key file (%s) does not exist", privKeyFilePath)
-
 	case err == nil || os.IsExist(err):
-		// private key file exists
+		// private key file already exists
+		return fmt.Errorf("private key file (%s) already exists", privKeyFilePath)
+
+	case os.IsNotExist(err):
+		// private key file does not exist, create a new one
 
 	default:
 		return fmt.Errorf("unable to check private key file (%s): %w", privKeyFilePath, err)
 	}
 
-	privKey, err := p2p.ReadEd25519PrivateKeyFromPEMFile(privKeyFilePath)
+	// create identity
+	privateKey, publicKey, err := crypto.GenerateKeyPair(crypto.Ed25519, -1)
 	if err != nil {
-		return fmt.Errorf("reading private key file for peer identity failed: %w", err)
+		return fmt.Errorf("unable to generate Ed25519 private key for peer identity: %w", err)
 	}
 
-	peerID, err := peer.IDFromPublicKey(privKey.GetPublic())
+	// obtain Peer ID from public key
+	peerID, err := peer.IDFromPublicKey(publicKey)
 	if err != nil {
 		return fmt.Errorf("unable to get peer identity from public key: %w", err)
 	}
 
-	privKeyBytes, err := privKey.Raw()
+	privKeyBytes, err := privateKey.Raw()
 	if err != nil {
 		return fmt.Errorf("unable to get raw private key bytes: %w", err)
 	}
 
-	pubKeyBytes, err := privKey.GetPublic().Raw()
+	pubKeyBytes, err := publicKey.Raw()
 	if err != nil {
 		return fmt.Errorf("unable to get raw public key bytes: %w", err)
+	}
+
+	if err := p2p.WriteEd25519PrivateKeyToPEMFile(privKeyFilePath, privateKey); err != nil {
+		return fmt.Errorf("writing private key file for peer identity failed: %w", err)
 	}
 
 	fmt.Println("Your p2p private key (hex):   ", hex.EncodeToString(privKeyBytes))